Modules/requirements.txt
requests
colorama
lxml
bs4
shodan
requests_html
fake-useragent
python-whois
phonenumbers